+++ src/speech/cancellation_details.rs
@@ -43,8 +43,8 @@ impl CancellationDetails {
                 .get_property(PropertyId::CancellationDetailsReasonDetailedText, "")?;
 
             Ok(CancellationDetails {
-                reason: CancellationReason::from_u32(reason.assume_init()),
-                error_code: CancellationErrorCode::from_u32(error_code.assume_init()),
+                reason: CancellationReason::from_u32(reason.assume_init() as u32),
+                error_code: CancellationErrorCode::from_u32(error_code.assume_init() as u32),
                 error_details,
             })
         }
